Added ExistsInArena implementation to BasicTower. Changed type of attribute 'speed' to double.

+++ src/main/java/project/Projectile.java
@@ -13,7 +13,7 @@ public class Projectile implements Arena.MovesInArena {
     private Coordinates coordinates;
 
     // Stats
-    private int speed;
+    private double speed;
 
     // Target to pursue
     private Monster target;
